requirements.txt

crayons==0.4.0
pandas==0.25.1
SQLAlchemy==1.4.20
PyMySQL==0.9.3
mysql-connector-python==8.0.15
mysqlclient==1.4.2.post1
google-api-core==1.31.0
google-api-python-client==2.12.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
urllib3==1.26.6